refactor: 프로젝트 관련 컴포넌트를 project 폴더로 통합
- project-view.tsx, project-form.tsx 이동
- gantt-chart-view.tsx, kanban-board.tsx, task-detail-panel.tsx 이동
- 모든 import 경로 업데이트 (@/components/project/로 변경)
