+++ app/(dashboard)/project/[id]/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState, useEffect } from "react"
 import { useParams, useRouter } from "next/navigation"
-import { ProjectView } from "@/components/project-view"
+import { ProjectView } from "@/components/project/project-view"
 import { getProject } from "@/lib/storage"
 import { mockProjects, mockProjectTasks } from "@/lib/mock-data"
