+++ components/project/project-view.tsx
@@ -4,9 +4,9 @@ import { useState, useCallback, useEffect } from "react"
 import { BarChart3, Kanban, Table } from "lucide-react"
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { HierarchicalWBSTable } from "@/components/hierarchical-wbs-table"
-import { GanttChartView } from "@/components/gantt-chart-view"
-import { KanbanBoard } from "@/components/kanban-board"
-import { TaskDetailPanel } from "@/components/task-detail-panel"
+import { GanttChartView } from "@/components/project/gantt-chart-view"
+import { KanbanBoard } from "@/components/project/kanban-board"
+import { TaskDetailPanel } from "@/components/project/task-detail-panel"
 import { ProjectHeader } from "@/components/project/project-header"
 import { ViewSelector } from "@/components/project/view-selector"
 import { useProjectData } from "@/hooks/use-project-data"
